requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# Web3 and Blockchain
web3==6.11.4
eth-account==0.10.0
eth-utils==4.0.0
hexbytes==0.3.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
motor==3.3.2  # MongoDB async driver

# Machine Learning
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
torch==2.1.2
transformers==4.36.2
optuna==3.5.0
shap==0.44.0  # Model interpretability

# Data Processing
scipy==1.11.4
statsmodels==0.14.1
ta-lib==0.4.28  # Technical indicators
pykalman==0.9.7  # Kalman filters

# Monitoring and Logging
prometheus-client==0.19.0
grafana-api==1.0.3
loguru==0.7.2
sentry-sdk==1.39.2

# Notifications
python-telegram-bot==20.7
discord.py==2.3.2
sendgrid==6.11.0

# API Integrations
ccxt==4.2.15  # Cryptocurrency exchange library
pycoingecko==3.1.0

# Security
cryptography==41.0.7
pynacl==1.5.0
python-jose==3.3.0

# Web Framework (for dashboard/API)
fastapi==0.108.0
uvicorn==0.25.0
websockets==12.0
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
httpx==0.26.0
tenacity==8.2.3  # Retry logic
cachetools==5.3.2
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Performance
numba==0.58.1  # JIT compilation
cython==3.0.8
uvloop==0.19.0  # Fast event loop

# Data Validation
pydantic==2.5.3
marshmallow==3.20.1

# Blockchain Analysis
evmdasm==0.1.10  # EVM disassembler
py-solc-x==2.0.2  # Solidity compiler

# Additional ML Libraries
catboost==1.2.2
tensorflow==2.15.0  # Alternative to PyTorch
keras==2.15.0

# Time Series Analysis
prophet==1.1.5
arch==6.2.0  # ARCH/GARCH models

# Optimization
scipy==1.11.4
cvxpy==1.4.1  # Convex optimization

# Parallel Processing
joblib==1.3.2
multiprocess==0.70.15
ray==2.9.0  # Distributed computing

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# Rate Limiting
aioredis==2.0.1
aiocache==0.12.2

# Blockchain Monitoring
web3-ethereum-defi==0.24.0
eth-abi==4.2.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Code Quality
pylint==3.0.3
autopep8==2.0.4
isort==5.13.2

# Deployment
gunicorn==21.2.0
supervisor==4.2.5
fabric==3.2.2

# Environment Management
python-decouple==3.8

# GraphQL (if needed)
gql==3.5.0
graphene==3.3

# Additional Security
pyotp==2.9.0  # 2FA
argon2-cffi==23.1.0  # Password hashing

# Profiling and Debugging
memory-profiler==0.61.0
line-profiler==4.1.2
py-spy==0.3.14

# Version Control
gitpython==3.1.41